0.3.2分支

+++ arius-admin/arius-admin-common/src/main/java/com/didichuxing/datachannel/arius/admin/common/bean/entity/stats/ESClusterStatsCells.java
@@ -12,171 +12,139 @@ public class ESClusterStatsCells {
      * 集群名称，all则表示全集群信息
      */
     private String clusterName;
-
     /**
      * 集群状态
      */
     private String status;
-
     /**
      * 集群状态
      */
-    private int    statusType;
-
+    private Integer statusType;
     /**
      * 未分配shard
      */
-    private long   unAssignedShards;
-
+    private Long unAssignedShards;
     /**
      * pengdingtask梳理
      */
-    private long   numberPendingTasks;
-
+    private Long numberPendingTasks;
     /**
      * dataNode数量
      */
-    private long   numberDataNodes;
-
-    private long   numberMasterNodes;
-    private long   numberClientNodes;
-    private long   numberIngestNodes;
-
+    private Long numberDataNodes;
+    private Long numberMasterNodes;
+    private Long numberClientNodes;
+    private Long numberIngestNodes;
     /**
      * 集群节点数量
      */
-    private long   numberNodes;
-
+    private Long numberNodes;
     /**
      * 节点存活率
      */
-    private int    alivePercent;
-
+    private Integer alivePercent;
     /**
      * 集群重要等级
      */
-    private int    level;
-
+    private Integer level;
     /**
-     * 集群已使用容量，单位
+     * 集群已使用容量，单位bytes
      */
     private Long storeSize;
-
     /**
-     * 集群总容量，单位
+     * 集群总容量，单位bytes
      */
     private Long totalStoreSize;
-
     /**
-     * 集群空余容量，单位
+     * 集群空余容量，单位bytes
      */
     private Long freeStoreSize;
-
     /**
-     * 索引容量，单位
+     * 索引容量，单位bytes
      */
-    private double indexStoreSize;
-
+    private Long indexStoreSize;
     /**
      * 集群索引数
      */
-    private double totalIndicesNu;
-
+    private Long totalIndicesNu;
     /**
      * 集群模板数
      */
-    private int    totalTemplateNu;
-
+    private Integer totalTemplateNu;
     /**
      * 集群文档数
      */
-    private long   totalDocNu;
-
+    private Long totalDocNu;
     /**
      * 集群shard数量
      */
-    private long   shardNu;
-
+    private Long shardNu;
     /**
      * 每秒接受流量
      */
-    private double recvTransSize;
-
+    private Double recvTransSize;
     /**
      * 每秒发送流量
      */
-    private double sendTransSize;
-
+    private Double sendTransSize;
     /**
      * 集群写入tps
      */
-    private double writeTps;
-
+    private Double writeTps;
     /**
      * 集群读取tps
      */
-    private double readTps;
-
+    private Double readTps;
     /**
      * cpu相关
      */
-    private double cpuUsage;
-    private double cpuLoad1M;
-    private double cpuLoad5M;
-    private double cpuLoad15M;
-
+    private Double cpuUsage;
+    private Double cpuLoad1M;
+    private Double cpuLoad5M;
+    private Double cpuLoad15M;
     /**
      * 磁盘使用率
      */
-    private double diskUsage;
-
+    private Double diskUsage;
     /**
      * 集群总数
      */
-    private int    clusterNu;
-
+    private Integer clusterNu;
     /**
      * es节点数
      */
-    private double esNodeNu;
-
+    private Long esNodeNu;
     /**
      * 应用数量
      */
-    private int    appNu;
-
+    private Integer appNu;
     /**
      * 每日查询总量
      */
-    private long   queryTimesPreDay;
+    private Long queryTimesPreDay;
     /**
      * 集群稳定性
      */
-    private double sla;
-
-    private long   memUsed;
-    private long   memFree;
-    private long   memTotal;
-    private double memUsedPercent;
-    private double memFreePercent;
-
+    private Double sla;
+    private Long memUsed;
+    private Long memFree;
+    private Long memTotal;
+    private Long memUsedPercent;
+    private Long memFreePercent;
     /**
      * 查询耗时
      */
-    private double searchLatency;
-
+    private Double searchLatency;
     /**
      * 写入耗时
      */
-    private double indexingLatency;
-
+    private Double indexingLatency;
     /**
      * 任务耗时
      */
     private double taskCost;
-
     /**
      * 任务数量
      */
-    private long   taskCount;
+    private long taskCount;
 }
